+++ src/camera/camera.hpp
@@ -11,7 +11,7 @@ class camera {
 		camera(const float& w, const float& h, const float scr[4], const float& nr, const float& fr,
 				const float& fov, const point3& pos, const point3& look, const vec3& up);
 
-		void getRay(const int& x, const int& y, ray& r);
+		void getRay(const float& x, const float& y, ray& r);
 
 	private:
 		float screen[4];
